+++ applications/admin.py
@@ -81,9 +81,9 @@ class DraftApplicationAdmin(admin.ModelAdmin):
 
 admin.site.register(models.HackerApplication, admin_class=ApplicationAdmin)
 admin.site.register(models.VolunteerApplication, admin_class=OtherApplicationAdmin)
-admin.site.register(models.MentorApplication, admin_class=OtherApplicationAdmin)
-admin.site.register(models.SponsorApplication, admin_class=SponsorApplicationAdmin)
-admin.site.register(models.DraftApplication, admin_class=DraftApplicationAdmin)
+admin.site.register(models.mentor.MentorApplication, admin_class=OtherApplicationAdmin)
+admin.site.register(models.sponsor.SponsorApplication, admin_class=SponsorApplicationAdmin)
+admin.site.register(models.draft.DraftApplication, admin_class=DraftApplicationAdmin)
 admin.site.site_header = '%s Admin' % settings.HACKATHON_NAME
 admin.site.site_title = '%s Admin' % settings.HACKATHON_NAME
 admin.site.index_title = 'Home'
