+++ applications/forms/__init__.py
@@ -1,0 +1,5 @@
+from .base import *
+from .hacker import *
+from .mentor import *
+from .sponsor import *
+from .volunteer import *
